Restore keys from the oldCatalog if missing in newCatalog
The old behaviour was to restore only if the newCatalog had an empty string. We want to restore also if the key is missing. Use case: add a key, remove the key, add it back. When the key is added back, it is missing and we still want to restore it.