Use utf-8 as default encoding to read files
The encoding set on the transform is technically supposed to be used for the chunks going through the transform itself, and it's not necessarily set for a Transform, especially in object mode.  This falls back on `utf-8` as the encoding when loading a file from disk.

I suspect this should just always use `utf-8` always but I can't say for sure there isn't a use case for using the `encoding` parameter.